Move new dlc generator to original file and name
