+++ weis/dlc_driver/dlc_generator.py
@@ -1,8 +1,82 @@
 import numpy as np
 import os
+import logging
 import weis.inputs as sch
 from weis.dlc_driver.turbulence_models import IEC_TurbulenceModels
+from weis.aeroelasticse.CaseGen_General import CaseGen_General
+from weis.aeroelasticse.FileTools import remove_numpy
+from weis.aeroelasticse.utils import OLAFParams
 
+logger = logging.getLogger("wisdem/weis")
+
+
+
+# TODO: not sure where this should live, so it's a global for now
+# Could it be an input yaml?
+openfast_input_map = {
+    # Generic name: OpenFAST input (list if necessary)
+    'total_time': ("Fst","TMax"),
+    'transient_time': ("Fst","TStart"),
+    
+    'WindFile_type': ("InflowWind","WindType"),
+    'wind_speeds': ("InflowWind","HWindSpeed"),
+    'WindFile_name': ("InflowWind","FileName_BTS"),
+    'WindFile_name': ("InflowWind","Filename_Uni"),
+    'rotorD': ("InflowWind","RefLength"),
+    'WindHd': ("InflowWind","PropagationDir"),
+    'hub_height': ("InflowWind","RefHt_Uni"),
+    
+    'rot_speed_initial': ("ElastoDyn","RotSpeed"),
+    'pitch_initial': [("ElastoDyn","BlPitch1"),("ElastoDyn","BlPitch2"),("ElastoDyn","BlPitch3")],
+    'azimuth_init': ("ElastoDyn","Azimuth"),
+    'yaw_misalign': ("ElastoDyn","NacYaw"),
+    
+    'wave_Hs': ("HydroDyn","WaveHs"),
+    'wave_Tp': ("HydroDyn","WaveTp"),
+    'WaveHd': ("HydroDyn","WaveDir"),
+    'WaveGamma': ("HydroDyn","WavePkShp"),
+    'WaveSeed1': ("HydroDyn","WaveSeed1"),
+
+    'wave_model': ("HydroDyn","WaveMod"),
+    
+    'shutdown_time': [
+        ("ServoDyn","TPitManS1"),
+        ("ServoDyn","TPitManS2"),
+        ("ServoDyn","TPitManS3"),
+        ("ServoDyn","TimGenOf"),
+        ],
+
+    'startup_time': [
+        ("ServoDyn","TimGenOn"),
+        ("ServoDyn","TPCOn"),
+    ],
+        
+
+    'final_blade_pitch': [
+        ("ServoDyn","BlPitchF(1)"),
+        ("ServoDyn","BlPitchF(2)"),
+        ("ServoDyn","BlPitchF(3)"),
+        
+    ],
+
+    'fault_time': [("ServoDyn","TPitManS1"),("ServoDyn","TPitManS2"),("ServoDyn","TPitManS3")],
+    
+    'aero_mod': ("AeroDyn15","AFAeroMod"),
+    'wake_mod': ("AeroDyn15","WakeMod"),
+    'tau1_const': ("AeroDyn15","tau1_const"),
+
+
+    # 'dlc_label': ("DLC","Label"),
+    # 'wind_seed': ("DLC","WindSeed"),
+    # 'wind_speeds': ("DLC","MeanWS"),
+
+    # TODO: where should turbsim live?
+    # These aren't actually used to generate turbsim, the generic inputs are used
+    # However, I think it's better to be over-thorough and check that inputs are applied than the uncertainty of not checking any
+    'rand_seeds': ("TurbSim", "RandSeed1"),
+    'direction': ("TurbSim", "direction_pn"),
+    'shear': ("TurbSim", "shear_hv")
+}
 
 class DLCInstance(object):
 
@@ -41,9 +115,47 @@ class DLCInstance(object):
             setattr(self, key, options['turbulent_wind'][key])
 
 class DLCGenerator(object):
+    OF_dlccaseinputvars = {("InflowWind","WindType"):0,
+        ("ElastoDyn","RotSpeed"):0,
+        ("ElastoDyn","BlPitch1"):90.0,
+        ("ElastoDyn","BlPitch2"):90.0,
+        ("ElastoDyn","BlPitch3"):90.0,
+        ("ServoDyn","TPitManS1"):9999,
+        ("ElastoDyn","Azimuth"):0,
+        ("InflowWind","PropagationDir"):0,
+        ("HydroDyn","WaveHs"):0,
+        ("HydroDyn","WaveTp"):0,
+        ("HydroDyn","WaveDir"):0,
+        ("HydroDyn","WavePkShp"):0,
+        ("HydroDyn","WaveSeed1"):0,
+        ("Fst","TMax"):0,
+        ("Fst","TStart"):0,
+        ("DLC","Label"):0,
+        ("DLC","WindSeed"):0,
+        ("DLC","MeanWS"):0,
+        ("ElastoDyn","NacYaw"):0,
+        ("AeroDyn15","tau1_const"):0,
+        ("AeroDyn15","OLAF","DTfvw"):0,
+        ("AeroDyn15","OLAF","nNWPanels"):0,
+        ("AeroDyn15","OLAF","nNWPanelsFree"):0,
+        ("AeroDyn15","OLAF","nFWPanels"):0,
+        ("AeroDyn15","OLAF","nFWPanelsFree"):0,
+    }
+    
 
-    def __init__(self, ws_cut_in=4.0, ws_cut_out=25.0, ws_rated=10.0, wind_speed_class = 'I',
-                wind_turbulence_class = 'B', fix_wind_seeds=True, fix_wave_seeds=True, metocean={}):
+    def __init__(
+            self, 
+            ws_cut_in=4.0, 
+            ws_cut_out=25.0, 
+            ws_rated=10.0, 
+            wind_speed_class = 'I',
+            wind_turbulence_class = 'B', 
+            fix_wind_seeds=True, 
+            fix_wave_seeds=True, 
+            metocean={},
+            initial_condition_table = {},
+            default_options = {},
+            ):
         self.ws_cut_in = ws_cut_in
         self.ws_cut_out = ws_cut_out
         self.wind_speed_class = wind_speed_class
@@ -60,6 +172,16 @@ class DLCGenerator(object):
             self.rng_wave = np.random.default_rng()
         self.n_cases = 0
         self.n_ws_dlc11 = 0
+
+        # Set and update default_options, applied to dlc_options and first group in case_inputs
+        self.default_options = {
+            'wake_mod': 1,
+            'wave_model': 2,
+        }
+        self.default_options.update(default_options)
+
+        # Init openfast case list
+        self.openfast_case_inputs = []
 
         # Metocean conditions
         self.mo_ws = metocean['wind_speed']
@@ -88,6 +210,8 @@ class DLCGenerator(object):
         self.wave_Hs1 = np.array([metocean['wave_height1']])
         self.wave_Tp1 = np.array([metocean['wave_period1']])
 
+        self.initial_condition_table = initial_condition_table
+
     def IECwind(self):
         self.IECturb = IEC_TurbulenceModels()
         self.IECturb.Turbine_Class = self.wind_speed_class
@@ -111,16 +235,12 @@ class DLCGenerator(object):
         return wind_speeds
 
     def get_wind_seeds(self, options, wind_speeds):
-        if 'yaw_misalign' in options:
-            n_yaw_ms = len(options['yaw_misalign'])
-        else:
-            n_yaw_ms = 1
         
         if len(options['wind_seed']) > 0:
             wind_seeds = np.array( [int(m) for m in options['wind_seed']] )
         else:
-            wind_seeds = self.rng_wind.integers(2147483648, size=options['n_seeds']*len(wind_speeds) * n_yaw_ms * options['n_azimuth'], dtype=int)
-            wind_speeds = np.repeat(wind_speeds, options['n_seeds'] * n_yaw_ms * options['n_azimuth'])
+            wind_seeds = self.rng_wind.integers(2147483648, size=options['n_seeds']*len(wind_speeds), dtype=int)
+            wind_speeds = np.repeat(wind_speeds, options['n_seeds'])
 
         return wind_speeds, wind_seeds
 
@@ -203,11 +323,27 @@ class DLCGenerator(object):
             raise Exception("The vector of probabilities must have either length=1 or the same length of wind speeds")
         if abs(sum(probabilities) - 1.) > 1.e-3:
             raise Exception("The vector of probabilities must sum to 1")
+        
+        metocean_case_info = {}
+        metocean_case_info['wind_speeds'] = wind_speeds
+        metocean_case_info['rand_seeds'] = wind_seeds
+        metocean_case_info['wave_seeds'] = wave_seeds
+        metocean_case_info['wind_heading'] = wind_heading
+        metocean_case_info['wave_Hs'] = wave_Hs
+        metocean_case_info['wave_Tp'] = wave_Tp
+        # metocean_case_info['current_speeds'] = current_speeds
+        metocean_case_info['wave_gamma'] = wave_gamma
+        metocean_case_info['wave_heading'] = wave_heading
+        metocean_case_info['probabilities'] = probabilities       
+        # metocean_case_info['current_std'] = self.mo_current_std       
+        
+        return metocean_case_info
 
-        return wind_speeds, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, probabilities
 
     def generate(self, label, options):
-        known_dlcs = [1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 5.1, 6.1, 6.3, 6.4, 6.5, 12.1]
+        # Use schema to determine known_dlcs (weis/inputs/modeling_schema.yaml)
+        known_dlcs = sch.validation.get_modeling_schema()['properties']['DLC_driver']['properties']['DLCs']['items']['properties']['DLC']['enum']
+        self.OF_dlccaseinputs = {key: None for key in known_dlcs}
 
         # Get extreme wind speeds
         self.IECwind()
@@ -226,667 +362,646 @@ class DLCGenerator(object):
 
         self.n_cases = len(self.cases)
 
-    def generate_custom(self, options):
-        pass
+    def generate_cases(self,generic_case_inputs,dlc_options):
+        '''
+        This method will generate the simulation inputs for each design load case
 
-    def generate_1p1(self, options):
-        # Power production normal turbulence model - ultimate loads
-        wind_speeds, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the normal sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_NSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_NSS)
-        # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
-        else: # default
-            yaw_misalign = [0]
-        yaw_misalign_deg = np.array(yaw_misalign * int(len(wind_speeds) / len(yaw_misalign)))
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.URef = ws
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.yaw_misalign = yaw_misalign_deg[i_WiSe]
-            idlc.turbulent_wind = True
-            idlc.label = '1.1'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            idlc.PSF = 1.2 * 1.25
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+        generic_case_inputs is a list of lists of strings with the options used to create a case matrix
+        dlc_options is a dictionary, some of its keys will be in generic_case_inputs and used to generate the cases
+            Other keys include:
+            sea_state is a string: either severe or normal
+            label is the string label used in when the wind inputs are created
+        '''
 
-        self.n_ws_dlc11 = len(np.unique(wind_speeds))
+        # Handle default options
+        if 'sea_state' not in dlc_options:
+            dlc_options['sea_state'] = 'normal'
 
-    def generate_1p2(self, options):
-        # Power production normal turbulence model - fatigue loads
-        wind_speeds, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, probabilities = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the fatigue analysis
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_F)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_F)
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.URef = ws
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.probability = probabilities[i_WaH]
-            idlc.turbulent_wind = True
-            idlc.label = '1.2'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            idlc.PSF = 1.
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+        if 'PSF' not in dlc_options:
+            dlc_options['PSF'] = 1.35
 
-    def generate_1p3(self, options):
-        # Power production extreme turbulence model - ultimate loads
-        wind_speeds, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the normal sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_NSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_NSS)
-        # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
-        else: # default
-            yaw_misalign = [0]
-        yaw_misalign_deg = np.array(yaw_misalign * int(len(wind_speeds) / len(yaw_misalign)))
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.URef = ws
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.yaw_misalign = yaw_misalign_deg[i_WiSe]
-            idlc.IEC_WindType = self.wind_speed_class_num + 'ETM'
-            idlc.turbulent_wind = True
-            idlc.label = '1.3'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+       
+        # Generate case list, both generic and OpenFAST specific
+        self.set_time_options(dlc_options)
+        met_options = self.gen_met_options(dlc_options, sea_state=dlc_options['sea_state'])
+        generic_case_list = self.apply_initial_conditions(generic_case_inputs,dlc_options, met_options)
+        generic_case_list = self.gen_case_list(dlc_options,met_options,generic_case_inputs)
 
-    def generate_1p4(self, options):
-        # Extreme coherent gust with direction change - ultimate loads
-        wind_speeds, _, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        directions = ['n', 'p']
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the normal sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_NSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_NSS)
-        # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
-        else: # default
-            yaw_misalign = [0]
-        yaw_misalign_deg = np.array(yaw_misalign * int(len(wind_speeds) / len(yaw_misalign)))
-        # Set azimuth start positions, tile so length is same as wind_seeds
-        azimuth_inits = np.tile(
-            np.linspace(0.,120.,options['n_azimuth'],endpoint=False),
-            int(len(wind_speeds)/options['n_azimuth'])
-            )
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            for direction in directions:
-                idlc = DLCInstance(options=options)
-                idlc.URef = ws
-                idlc.wave_seed1 = wave_seeds[i_WaSe]
-                idlc.wind_heading = wind_heading[i_WiH]
-                idlc.wave_height = wave_Hs[i_Hs]
-                idlc.wave_period = wave_Tp[i_Tp]
-                idlc.wave_gamma = wave_gamma[i_WG]
-                idlc.wave_heading = wave_heading[i_WaH]
-                idlc.yaw_misalign = yaw_misalign_deg[i_WaSe]
-                idlc.azimuth_init = azimuth_inits[i_WaSe]
-                idlc.IEC_WindType = 'ECD'
+        # DLC specific: Make idlc for other parts of WEIS (mostly turbsim generation)
+        for _, case in enumerate(generic_case_list):
+            idlc = DLCInstance(options=dlc_options)
+            if dlc_options['IEC_WindType'] == 'ECD':
                 idlc.turbulent_wind = False
-                idlc.label = '1.4'
-                if options['analysis_time'] > 0:
-                    idlc.analysis_time = options['analysis_time']
-                if options['transient_time'] >= 0:
-                    idlc.transient_time = options['transient_time']
-                idlc.direction_pn = direction
-                self.cases.append(idlc)
-                if len(wind_heading)>1:
-                    i_WiH+=1
-                if len(wave_gamma)>1:
-                    i_WG+=1
-                if len(wave_heading)>1:
-                    i_WaH+=1
-            # Same wave height, period, and seed per direction, check whether this is allowed or change seed sampling
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-
-    def generate_1p5(self, options):
-        # Extreme wind shear - ultimate loads
-        wind_speeds, _, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the normal sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_NSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_NSS)
-        # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
-        else: # default
-            yaw_misalign = [0]
-        yaw_misalign_deg = np.array(yaw_misalign * int(len(wind_speeds) / len(yaw_misalign)))
-        directions = ['p', 'n']
-        shears=['h', 'v']
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            for direction in directions:
-                for shear in shears:
-                    idlc = DLCInstance(options=options)
-                    idlc.URef = ws
-                    idlc.wave_seed1 = wave_seeds[i_WaSe]
-                    idlc.wind_heading = wind_heading[i_WiH]
-                    idlc.wave_height = wave_Hs[i_Hs]
-                    idlc.wave_period = wave_Tp[i_Tp]
-                    idlc.wave_gamma = wave_gamma[i_WG]
-                    idlc.wave_heading = wave_heading[i_WaH]
-                    idlc.yaw_misalign = yaw_misalign_deg[i_WaSe]
-                    idlc.IEC_WindType = 'EWS'
-                    idlc.turbulent_wind = False
-                    idlc.label = '1.5'
-                    if options['analysis_time'] > 0:
-                        idlc.analysis_time = options['analysis_time']
-                    if options['transient_time'] >= 0:
-                        idlc.transient_time = options['transient_time']
-                    idlc.sigma1 = self.IECturb.NTM(ws)
-                    idlc.direction_pn = direction
-                    idlc.shear_hv = shear
-                    self.cases.append(idlc)
-                    if len(wind_heading)>1:
-                        i_WiH+=1
-                    if len(wave_gamma)>1:
-                        i_WG+=1
-                    if len(wave_heading)>1:
-                        i_WaH+=1
-            # Same wave height, period, and seed per direction, check whether this is allowed or change seed sampling
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-
-    def generate_1p6(self, options):
-        # Power production normal turbulence model - severe sea state
-        wind_speeds, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the severe sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_SSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_SSS)
-        # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
-        else: # default
-            yaw_misalign = [0]
-        yaw_misalign_deg = np.array(yaw_misalign * int(len(wind_speeds) / len(yaw_misalign)))
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.URef = ws
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.yaw_misalign = yaw_misalign_deg[i_WiSe]
-            idlc.turbulent_wind = True
-            idlc.label = '1.6'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
-
-    def generate_5p1(self, options):
-        # Power production normal turbulence model - severe sea state
-        wind_speeds, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the normal sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_NSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_NSS)
-        # Set azimuth start positions, tile so length is same as wind_seeds
-        azimuth_inits = np.tile(
-            np.linspace(0.,120.,options['n_azimuth'],endpoint=False),
-            int(len(wind_seeds)/options['n_azimuth'])
-            )
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.URef = ws
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.azimuth_init = azimuth_inits[i_WiSe]
-            idlc.turbulent_wind = True
-            idlc.label = '5.1'
-            idlc.turbine_status = 'operating-shutdown'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            if options['shutdown_time'] > options['analysis_time']:
-                raise Exception(f"DLC 5.1 was selected, but the shutdown_time ({options['shutdown_time']}) option is greater than the analysis_time ({options['analysis_time']})")
+                idlc.direction_pn = case['direction']
+            elif dlc_options['IEC_WindType'] == 'EWS':
+                idlc.turbulent_wind = False
+                idlc.direction_pn = case['direction']
+                idlc.shear_hv = case['shear']
+                idlc.sigma1 = self.IECturb.NTM(case['wind_speeds'])
             else:
-                idlc.shutdown_time = options['shutdown_time']
+                idlc.turbulent_wind = True
+                idlc.RandSeed1 = case['rand_seeds']
+            idlc.URef = case['wind_speeds']
+            idlc.label = dlc_options['label']
+            idlc.total_time = case['total_time']
+            idlc.IEC_WindType = dlc_options['IEC_WindType']
+            if dlc_options['label'] == '1.2':
+                idlc.probability = probabilities[i_WaH]
             self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+            
 
-    def generate_6p1(self, options):
-        # Parked (standing still or idling) - extreme wind model 50-year return period - ultimate loads
-        options['wind_speed'] = [50,50]  # set dummy, so wind seeds are correct
-        _, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
+    def apply_sea_state(self,met_options,sea_state='normal'):
+        '''
+        Apply waves based on the expected values provided in the metocean inputs
+        Will use met_options as an input and modify that dict
+        sea_state can be normal, severe
+        '''
+        allowed_sea_states = ['normal','severe','50-year','1-year']
+        if sea_state not in allowed_sea_states:
+            raise Exception(f'Selected sea state of {sea_state} is not in allowed_sea_states: {allowed_sea_states}')
+        
+        # Select wind speed, sea state lookup
+        if sea_state == 'normal':
+            wind_speed_table = self.mo_ws
+            wave_height_table = self.mo_Hs_NSS
+            wave_period_table = self.mo_Tp_NSS
+        elif sea_state == 'severe':
+            wind_speed_table = self.mo_ws
+            wave_height_table = self.mo_Hs_SSS
+            wave_period_table = self.mo_Tp_SSS
+        elif sea_state == '50-year':
+            wind_speed_table = [50.]
+            wave_height_table = self.wave_Hs50
+            wave_period_table = self.wave_Tp50
+        elif sea_state == '1-year':
+            wind_speed_table = [50.]
+            wave_height_table = self.wave_Hs50
+            wave_period_table = self.wave_Tp50
+
+
+        # If the user has not defined Hs (wave_height in modopts) and Tp (wave_period in modopts), apply the metocean conditions defined by the table
+        if len(met_options['wave_Hs'])==0:
+            met_options['wave_Hs'] = np.interp(met_options['wind_speeds'], wind_speed_table, wave_height_table)
+        if len(met_options['wave_Tp'])==0:
+            met_options['wave_Tp'] = np.interp(met_options['wind_speeds'], wind_speed_table, wave_period_table)
+
+    def set_time_options(self, options):
+        '''
+        Handle time options and add total_time to dict
+        Default for analysis and transient_time is 0
+        '''
+        if options['analysis_time'] > 0:
+            options['analysis_time'] = options['analysis_time']
+        else:
+            options['analysis_time'] = 600.
+        if options['transient_time'] >= 0:
+            options['transient_time'] = options['transient_time']
+        options['total_time'] = options['analysis_time'] + options['transient_time']
+
+    def gen_case_list(self,dlc_options, met_options, generic_case_inputs):
+        '''
+        Generate case list from generic_case_inputs
+        TODO: this whole thing could be moved into generate_cases, thoughts?
+        '''
+
+        
+        # Combine
+        comb_options = combine_options(dlc_options,met_options)
+
+        # Check that all inputs are valid options
+        all_inputs = sum(generic_case_inputs, [])
+        for input in all_inputs:
+            if not input in comb_options:
+                raise Exception(f'The desired input {input} is not defined. Options include {comb_options.keys()}')
+
+        # Setup generic cross product of inputs: 
+        gen_case_inputs = {}
+        for i_group, group in enumerate(generic_case_inputs):
+            first_array_len = len(comb_options[group[0]])
+            for input in group:
+                
+                # Check that all inputs are of equal length
+                if len(comb_options[input]) != first_array_len:
+                    raise Exception(f'The input options in group {i_group} are not equal.  This group contains: {group}')
+
+                gen_case_inputs[input] = {'vals': comb_options[input], 'group': i_group}
+            
+        # Generate generic case list
+        generic_case_list, _ = CaseGen_General(gen_case_inputs,save_matrix=False)
+
+        case_inputs_openfast = map_generic_to_openfast(generic_case_inputs, comb_options)
+        self.openfast_case_inputs.append(case_inputs_openfast)
+        return generic_case_list
+
+    def gen_met_options(self, dlc_options, sea_state='normal'):
+        '''
+        Determine metocean options based on dlcs and sea state requested
+        met_options includes wind, waves, seeds, etc.
+
+        TODO: what input conditions are required of self?
+        TODO: what is required in dlc_options?
+        '''
+        met_options = self.get_metocean(dlc_options)
+        
+        # Apply normal wave conditions based on wind speeds
+        self.apply_sea_state(met_options,sea_state=sea_state)
+        make_equal_length(met_options,'wind_speeds')
+        return met_options
+
+    def apply_initial_conditions(self,generic_case_inputs, dlc_options, met_options):
+        '''
+        Add available initial conditions to generic_case_inputs and interpolate options based on initial_condition_table
+        This is performed within each dlc generator function
+
+        '''
+        
+        # These allowed_ics should map to input in openfast_input_map
+        allowed_ics = ['pitch_initial','rot_speed_initial','tau1_const']
+
+        
+        if self.initial_condition_table and dlc_options['turbine_status'] == 'operating': # there is an IC table that's not empty
+            dlc_wind_speeds = met_options['wind_speeds']  # need to use met_options wind speeds because it accounts for seeds
+            # find group with wind_speed
+            wind_group = ['wind_speeds' in gci for gci in generic_case_inputs].index(True)
+            group = generic_case_inputs[wind_group]
+
+            for initial_condition in allowed_ics:
+                if initial_condition in self.initial_condition_table:
+                    group.append(initial_condition)
+                    dlc_options[initial_condition] = np.interp(dlc_wind_speeds,self.initial_condition_table['U'],self.initial_condition_table[initial_condition])
+                
+            # Apply new group
+            generic_case_inputs[wind_group] = group
+            
+        return generic_case_inputs
+
+    def generate_1p1(self, dlc_options):
+        # Power production normal turbulence model - normal sea state
+        
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Handle DLC Specific options:
+        dlc_options['label'] = '1.1'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['PSF'] = 1.35
+
         # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
         else: # default
-            yaw_misalign = [-8., 8.]
-        yaw_misalign_deg = np.array(yaw_misalign * options['n_seeds'])
-        if len(wave_Hs)==0:
-            wave_Hs = self.wave_Hs50
-        if len(wave_Tp)==0:
-            wave_Tp = self.wave_Tp50
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for yaw_ms in yaw_misalign_deg:
-            idlc = DLCInstance(options=options)
-            if idlc.URef < 0:   # default is -1, this allows us to set custom V_50
-                idlc.URef = self.V_e50
-            idlc.yaw_misalign = yaw_ms
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.IEC_WindType = self.wind_speed_class_num + 'EWM50'
-            idlc.turbulent_wind = True
-            if idlc.turbine_status == 'operating':
-                idlc.turbine_status = 'parked-still'
-            idlc.label = '6.1'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+            dlc_options['yaw_misalign'] = [0]
 
-    def generate_6p3(self, options):
-        # Parked (standing still or idling) - extreme wind model 1-year return period - ultimate loads
-        options['wind_speed'] = [50,50]  # set dummy, so wind seeds are correct
-        _, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+    def generate_1p2(self, dlc_options):
+        # Power production normal turbulence model - fatigue loads
+        
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Handle DLC Specific options:
+        dlc_options['label'] = '1.2'
+        dlc_options['sea_state'] = 'normal'
+
         # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
         else: # default
-            yaw_misalign = [-20., 20.]
-        yaw_misalign_deg = np.array(yaw_misalign * options['n_seeds'])
-        if len(wave_Hs)==0:
-            wave_Hs = self.wave_Hs1
-        if len(wave_Tp)==0:
-            wave_Tp = self.wave_Tp1
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for yaw_ms in yaw_misalign_deg:
-            idlc = DLCInstance(options=options)
-            if idlc.URef < 0:   # default is -1, this allows us to set custom V_50
-                idlc.URef = self.V_e1
-            idlc.yaw_misalign = yaw_ms
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.IEC_WindType = self.wind_speed_class_num + 'EWM1'
-            idlc.turbulent_wind = True
-            if idlc.turbine_status == 'operating':
-                idlc.turbine_status = 'parked-still'
-            idlc.label = '6.3'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+            dlc_options['yaw_misalign'] = [0]
 
-    def generate_6p4(self, options):
-        # Parked (standing still or idling) - normal turbulence model - fatigue loads
-        wind_speeds = np.arange(self.ws_cut_in, 0.7 * self.V_ref, options['ws_bin_size'])
-        if wind_speeds[-1] != self.V_ref:
-            wind_speeds = np.append(wind_speeds, self.V_ref)
-        wind_speeds, wind_seeds = self.get_wind_seeds(options, wind_speeds)
-        wind_speeds = np.repeat(wind_speeds, options['n_seeds'])
-        wave_seeds = self.get_wave_seeds(options, wind_speeds)
-        _, _, _, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        # If the user has not defined Hs and Tp, apply the metocean conditions for the normal sea state
-        if len(wave_Hs)==0:
-            wave_Hs = np.interp(wind_speeds, self.mo_ws, self.mo_Hs_NSS)
-        if len(wave_Tp)==0:
-            wave_Tp = np.interp(wind_speeds, self.mo_ws, self.mo_Tp_NSS)
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.URef = ws
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.turbulent_wind = True
-            if idlc.turbine_status == 'operating':
-                idlc.turbine_status = 'parked-still'
-            idlc.label = '6.4'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
 
-    def generate_6p5(self, options):
-        # Parked (standing still or idling) - extreme wind model 500-year return period - ultimate loads
-        options['wind_speed'] = [50,50]  # set dummy, so wind seeds are correct
-        _, wind_seeds, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+
+    def generate_1p3(self, dlc_options):
+        # Power production extreme turbulence model - ultimate loads
+        
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Handle DLC Specific options:
+        dlc_options['label'] = '1.1'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['IEC_WindType'] = '1ETM'
+
         # Set yaw_misalign, else default
-        if 'yaw_misalign' in options:
-            yaw_misalign = options['yaw_misalign']
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
         else: # default
-            yaw_misalign = [-8., 8.]
-        yaw_misalign_deg = np.array(yaw_misalign * options['n_seeds'])
-        if len(wave_Hs)==0:
-            wave_Hs = self.wave_Hs1
-        if len(wave_Tp)==0:
-            wave_Tp = self.wave_Tp1
-        # Counter for wind seed
-        i_WiSe=0
-        # Counters for wave conditions
-        i_WaSe=0
-        i_Hs=0
-        i_Tp=0
-        i_WiH=0
-        i_WG=0
-        i_WaH=0
-        for yaw_ms in yaw_misalign_deg:
-            idlc = DLCInstance(options=options)
-            if idlc.URef < 0:   # default is -1, this allows us to set custom V_50
-                idlc.URef = self.V_e50 * 1.125
-            idlc.yaw_misalign = yaw_ms
-            idlc.RandSeed1 = wind_seeds[i_WiSe]
-            idlc.wave_seed1 = wave_seeds[i_WaSe]
-            idlc.wind_heading = wind_heading[i_WiH]
-            idlc.wave_height = wave_Hs[i_Hs]
-            idlc.wave_period = wave_Tp[i_Tp]
-            idlc.wave_gamma = wave_gamma[i_WG]
-            idlc.wave_heading = wave_heading[i_WaH]
-            idlc.IEC_WindType = self.wind_speed_class_num + 'EWM1'
-            idlc.turbulent_wind = True
-            if idlc.turbine_status == 'operating':
-                idlc.turbine_status = 'parked-still'
-            idlc.label = '6.5'
-            if options['analysis_time'] > 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
-            self.cases.append(idlc)
-            if len(wind_seeds)>1:
-                i_WiSe+=1
-            if len(wave_seeds)>1:
-                i_WaSe+=1
-            if len(wind_heading)>1:
-                i_WiH+=1
-            if len(wave_Hs)>1:
-                i_Hs+=1
-            if len(wave_Tp)>1:
-                i_Tp+=1
-            if len(wave_gamma)>1:
-                i_WG+=1
-            if len(wave_heading)>1:
-                i_WaH+=1
+            dlc_options['yaw_misalign'] = [0]
+
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+    def generate_1p4(self, dlc_options):
+        # Extreme coherent gust with direction change - ultimate loads
+        
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Handle DLC Specific options:
+        dlc_options['label'] = '1.4'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['IEC_WindType'] = 'ECD'
+        dlc_options['direction'] = ['n', 'p']
+
+        # Set yaw_misalign, else default
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
+        else: # default
+            dlc_options['yaw_misalign'] = [0]
+        
+        dlc_options['azimuth_init'] = np.linspace(0.,120.,dlc_options['n_azimuth'],endpoint=False)
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign','azimuth_init']) # group 2: 
+        generic_case_inputs.append(['direction']) # group 3: 
+
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+        
+    def generate_1p5(self, dlc_options):
+        # Extreme wind shear - ultimate loads
+        
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Handle DLC Specific options:
+        dlc_options['label'] = '1.5'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['IEC_WindType'] = 'EWS'
+        dlc_options['direction'] = ['n', 'p']
+        dlc_options['shear'] = ['h', 'v']
+        
+
+        # Set yaw_misalign, else default
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
+        else: # default
+            dlc_options['yaw_misalign'] = [0]
+        
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2: 
+        generic_case_inputs.append(['direction']) # group 3: 
+        generic_case_inputs.append(['shear']) # group 4: 
+
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+    def generate_1p6(self, dlc_options):
+        # Power production normal turbulence model - severe sea state
+
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # DLC Specific options:
+        dlc_options['label'] = '1.6'
+        dlc_options['sea_state'] = 'severe'
+        dlc_options['IEC_WindType'] = 'NTM'
+
+        # Set yaw_misalign, else default
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
+        else: # default
+            dlc_options['yaw_misalign'] = [0]
+
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','wake_mod','wave_model'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+    def generate_3p1(self, dlc_options):
+        # Start up - normal wind - fatigue
+        # 
+        
+        # Get default options
+        dlc_options.update(self.default_options)      
+        
+        # DLC Specific options:
+        dlc_options['label'] = '3.1'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['IEC_WindType'] = 'NTM'
+        dlc_options['pitch_initial'] = 90.
+        dlc_options['turbine_status'] = 'parked-idling'     # initial turbine status is what matters here
+
+        # Specify startup time for this case
+        if dlc_options['startup_time'] > dlc_options['analysis_time']:
+            raise Exception(f"DLC 3.1 was selected, but the startup_time ({dlc_options['startup_time']}) option is greater than the analysis_time ({dlc_options['analysis_time']})")
+        else:
+            dlc_options['startup_time'] = dlc_options['startup_time']
+
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','startup_time','wake_mod','wave_model','pitch_initial'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        # generic_case_inputs.append(['azimuth_init']) # group 2
+      
+        self.generate_cases(generic_case_inputs,dlc_options)
+
     
-    def generate_12p1(self, options):
-        # Pass through uniform wind input
-        wind_speeds, _, wave_seeds, wind_heading, wave_Hs, wave_Tp, wave_gamma, wave_heading, _ = self.get_metocean(options)
-        for ws in wind_speeds:
-            idlc = DLCInstance(options=options)
-            idlc.label = '12.1'
-            idlc.IEC_WindType = 'Custom'
-            idlc.URef = wind_speeds
-            idlc.turbulent_wind = False
-            # idlc.wind_file = options['wind_file']
-            if options['analysis_time'] >= 0:
-                idlc.analysis_time = options['analysis_time']
-            if options['transient_time'] >= 0:
-                idlc.transient_time = options['transient_time']
+    def generate_5p1(self, dlc_options):
+        # Power production normal turbulence model - shutdown with varous azimuth initial conditions
+        # 
+        
+        # Get default options
+        dlc_options.update(self.default_options)      
+        
+        # DLC Specific options:
+        dlc_options['label'] = '5.1'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['IEC_WindType'] = 'NTM'
+        dlc_options['final_blade_pitch'] = 90.
 
-            self.cases.append(idlc)
+        
+        # azimuth starting positions
+        dlc_options['azimuth_init'] = np.linspace(0.,120.,dlc_options['n_azimuth'],endpoint=False)
+
+        # Specify shutdown time for this case
+        if dlc_options['shutdown_time'] > dlc_options['analysis_time']:
+            raise Exception(f"DLC 5.1 was selected, but the shutdown_time ({dlc_options['shutdown_time']}) option is greater than the analysis_time ({dlc_options['analysis_time']})")
+        else:
+            dlc_options['shutdown_time'] = dlc_options['shutdown_time']
+
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','shutdown_time','wake_mod','wave_model','final_blade_pitch'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['azimuth_init']) # group 2
+      
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+
+    def generate_6p1(self, dlc_options):
+        # Parked (standing still or idling) - extreme wind model 50-year return period - ultimate loads
+        # extra dlc_options: 
+        # yaw_misalign: default = [-8,8]
+
+        # Get default options
+        dlc_options.update(self.default_options)
+
+        # DLC Specific options:
+        dlc_options['label'] = '6.1'
+        dlc_options['sea_state'] = '50-year'
+        dlc_options['IEC_WindType'] = self.wind_speed_class_num + 'EWM50'
+
+        # yaw_misalign
+        if 'yaw_misalign' not in dlc_options:
+            dlc_options['yaw_misalign'] = [-8,8]
+
+        dlc_options['wind_speed'] = [50]   # placeholder, could be anything as long as the length is 1, since the EWM50 is just a single speed that turbsim will determine 
+
+        # parked options
+        dlc_options['turbine_status'] = 'parked-idling'
+        dlc_options['wake_mod'] = 0
+        dlc_options['pitch_initial'] = 90.
+        dlc_options['rot_speed_initial'] = 0.
+        dlc_options['shutdown_time'] = 0.
+        dlc_options['final_blade_pitch'] = 90.
+
+
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','wake_mod','wave_model','pitch_initial',
+                                    'rot_speed_initial','shutdown_time','final_blade_pitch'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+      
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+    def generate_6p2(self, dlc_options):
+        # Parked (standing still or idling) - extreme wind model 50-year return period - ultimate loads
+        # This is the same as DLC 6.1 in the 61400-3-1 standards, except there's a loss of electrical network.
+        # In DLC 6.1, the generator is disabled already, so the only difference in 6.2 may be that users may want to simulate larger yaw misalignments
+        # extra dlc_options: 
+        # yaw_misalign: default = [-8,8]
+
+        # Get default options
+        dlc_options.update(self.default_options)
+
+        # DLC Specific options:
+        dlc_options['label'] = '6.1'
+        dlc_options['sea_state'] = '50-year'
+        dlc_options['IEC_WindType'] = self.wind_speed_class_num + 'EWM50'
+
+        # yaw_misalign
+        if 'yaw_misalign' not in dlc_options:
+            dlc_options['yaw_misalign'] = [-8,8]
+
+        dlc_options['wind_speed'] = [50]   # placeholder, could be anything as long as the length is 1, since the EWM50 is just a single speed that turbsim will determine 
+
+        # parked options
+        dlc_options['turbine_status'] = 'parked-idling'
+        dlc_options['wake_mod'] = 0
+        dlc_options['pitch_initial'] = 90.
+        dlc_options['rot_speed_initial'] = 0.
+        dlc_options['shutdown_time'] = 0.
+        dlc_options['final_blade_pitch'] = 90.
+
+
+        # DLC-specific: define groups
+        # These options should be the same length and we will generate a matrix of all cases
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','wake_mod','wave_model','pitch_initial',
+                                    'rot_speed_initial','shutdown_time','final_blade_pitch'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+      
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+
+    def generate_6p3(self, dlc_options):
+        # Parked (standing still or idling) - extreme wind model 1-year return period - ultimate loads, usually larger (20 deg) yaw offset
+
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Set DLC Specific options:
+        # These three are required
+        dlc_options['label'] = '6.3'
+        dlc_options['sea_state'] = '1-year'
+        dlc_options['IEC_WindType'] = self.wind_speed_class_num + 'EWM1'
+
+        # Set dlc-specific options, like yaw_misalign, initial azimuth
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
+        else: # default
+            dlc_options['yaw_misalign'] = [-20.,20.]
+
+        dlc_options['wind_speed'] = [50]   # placeholder, could be anything as long as the length is 1, since the EWM50 is just a single speed that turbsim will determine 
+
+        # parked options
+        dlc_options['turbine_status'] = 'parked-idling'
+        dlc_options['wake_mod'] = 0
+        dlc_options['pitch_initial'] = 90.
+        dlc_options['rot_speed_initial'] = 0.
+        dlc_options['shutdown_time'] = 0.
+        dlc_options['final_blade_pitch'] = 90.
+
+        # DLC-specific: define groups
+        # Groups are dependent variables, the cases are a cross product of the independent groups
+        # The options in each group should have the same length
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','wake_mod','wave_model','pitch_initial',
+                                    'rot_speed_initial','shutdown_time','final_blade_pitch'])  # group 0, (usually constants) turbine variables, DT, aero_modeling        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+
+        # This function does the rest and generates the individual cases for each DLC
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+
+    def generate_6p4(self, dlc_options):
+        # Parked (standing still or idling) - normal turbulence model - fatigue loads
+
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Set DLC Specific options:
+        # These three are required
+        dlc_options['label'] = '6.4'
+        dlc_options['sea_state'] = 'normal'
+        dlc_options['IEC_WindType'] = 'NTM'
+
+        # Set dlc-specific options, like yaw_misalign, initial azimuth
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
+        else: # default
+            dlc_options['yaw_misalign'] = [-8.,8.]
+
+        # parked options
+        dlc_options['turbine_status'] = 'parked-idling'
+        dlc_options['wake_mod'] = 0
+        dlc_options['pitch_initial'] = 90.
+        dlc_options['rot_speed_initial'] = 0.
+        dlc_options['shutdown_time'] = 0.
+        dlc_options['final_blade_pitch'] = 90.
+
+        # DLC-specific: define groups
+        # Groups are dependent variables, the cases are a cross product of the independent groups
+        # The options in each group should have the same length
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','wake_mod','wave_model','pitch_initial',
+                                    'rot_speed_initial','shutdown_time','final_blade_pitch'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+
+        # This function does the rest and generates the individual cases for each DLC
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+
+    def generate_new_dlc(self,dlc_options):
+        # Describe the new design load case
+
+        # Get default options
+        dlc_options.update(self.default_options)   
+        
+        # Set DLC Specific options:
+        # These three are required
+        dlc_options['label'] = '1.6'
+        dlc_options['sea_state'] = 'severe'
+        dlc_options['IEC_WindType'] = 'NTM'
+
+        # Set dlc-specific options, like yaw_misalign, initial azimuth
+        if 'yaw_misalign' in dlc_options:
+            dlc_options['yaw_misalign'] = dlc_options['yaw_misalign']
+        else: # default
+            dlc_options['yaw_misalign'] = [0]
+
+        # DLC-specific: define groups
+        # Groups are dependent variables, the cases are a cross product of the independent groups
+        # The options in each group should have the same length
+        generic_case_inputs = []
+        generic_case_inputs.append(['total_time','transient_time','wake_mod','wave_model'])  # group 0, (usually constants) turbine variables, DT, aero_modeling
+        generic_case_inputs.append(['wind_speeds','wave_Hs','wave_Tp', 'rand_seeds']) # group 1, initial conditions will be added here, define some method that maps wind speed to ICs and add those variables to this group
+        generic_case_inputs.append(['yaw_misalign']) # group 2
+
+        # This function does the rest and generates the individual cases for each DLC
+        self.generate_cases(generic_case_inputs,dlc_options)
+
+def make_equal_length(option_dict,target_name):
+    '''
+    This function will set the length of all the option_dicts to that of option_dict[target_name] if it's a scalar
+    '''
+    target_len = len(option_dict[target_name])
+    for key in option_dict:
+        if len(option_dict[key]) == 1:
+            option_dict[key] = option_dict[key] * target_len
+
+def combine_options(*dicts):
+    """
+    Combine option dictionarys and do standard processing, 
+    like removing numpy and turning everything into lists for case_inputs
+    
+    Args:
+        *dicts: Variable number of dictionaries.
+        
+    Returns:
+        dict: Combined dictionary.
+    """
+    comb_options = {}
+    for d in dicts:
+        comb_options.update(d)
+
+    comb_options = remove_numpy(comb_options)
+    
+    # Make all options a list
+    for opt in comb_options:
+        if not isinstance(comb_options[opt], list):  # if not a list
+            comb_options[opt] = [comb_options[opt]]
+
+    return comb_options
+
+def map_generic_to_openfast(generic_case_inputs, comb_options):
+    case_inputs_openfast = {}
+    for i_group, generic_case_group in enumerate(generic_case_inputs):
+        for generic_input in generic_case_group:
+            
+            if generic_input not in openfast_input_map.keys():
+                raise Exception(f'The input {generic_input} does not map to an OpenFAST input key in openfast_input_map')
+
+            openfast_input = openfast_input_map[generic_input]
+
+            if type(openfast_input) == list:
+                # Apply to all list of openfast_inputs
+                for of_input in openfast_input:
+                    case_inputs_openfast[of_input] = {'vals': comb_options[generic_input], 'group': i_group}
+
+            else:
+                case_inputs_openfast[openfast_input] = {'vals': comb_options[generic_input], 'group': i_group}
+
+    return case_inputs_openfast
 
 
 if __name__ == "__main__":
@@ -917,5 +1032,15 @@ if __name__ == "__main__":
         DLCopt = DLCs[i_DLC]
         dlc_generator.generate(DLCopt['DLC'], DLCopt)
 
-    print(dlc_generator.cases[5].URef)
-    print(dlc_generator.n_cases)
+    # print(dlc_generator.cases[1].URef)
+    # print(dlc_generator.n_cases)
+
+    FAST_runDirectory = '/Users/dzalkind/Tools/WEIS-DLC/examples/05_IEA-3.4-130-RWT/outputs/05_DLC15_new_setup/openfast_runs'
+    FAST_InputFile = 'weis_job'
+
+    case_list_all = []
+    for case_inputs in dlc_generator.openfast_case_inputs:
+        case_list, case_name = CaseGen_General(case_inputs, FAST_runDirectory, FAST_InputFile)
+        print('here')
+
+
